bug 8710: CONTD....Introducing a new user role in cloudstack called RESOURCE_DOMAIN_ADMIN. The role would have all the domain_admin rights and the rights to list zone,pods,clusters and so on. More info in the bug

+++ server/src/com/cloud/agent/manager/AgentManagerImpl.java
@@ -795,6 +795,8 @@ public class AgentManagerImpl implements AgentManager, HandlerFactory, ResourceS
 		String hostTag = cmd.getHostTag();
 		Map<String, String>bareMetalParams = new HashMap<String, String>();
 		
+		dcId = _accountMgr.checkAccessAndSpecifyAuthority(UserContext.current().getCaller(), dcId);
+		
 		// this is for standalone option
 		if (clusterName == null && clusterId == null) {
 			clusterName = "Standalone-" + url;
@@ -1303,7 +1305,7 @@ public class AgentManagerImpl implements AgentManager, HandlerFactory, ResourceS
 			throw new InvalidParameterValueException("Host with id "
 					+ id.toString() + " doesn't exist");
 		}
-
+		_accountMgr.checkAccessAndSpecifyAuthority(UserContext.current().getCaller(), host.getDataCenterId());
 		return deleteHost(id);
 	}
